requirements.txt
appdirs==1.4.3
cycler==0.10.0
Cython==0.25.2
gdbn==0.1
gnumpy==0.2
joblib==0.11
Kivy==1.9.1
Kivy-Garden==0.1.4
Lasagne==0.1
nolearn==0.6.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
pygame==1.9.3
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
scikit-learn==0.18.1
six==1.10.0
tabulate==0.7.7
Theano==0.9.0
